added ability to change name and zone number of watering zones
