+++ lib/grass_farmer_web/components/zones.ex
@@ -3,7 +3,7 @@ defmodule GrassFarmerWeb.Components.Zones do
 
   import GrassFarmerWeb.Components.StyleBlocks
 
-  alias GrassFarmer.Zone
+  alias GrassFarmer.{ Zone, PersistenceAdapter }
 
   @impl true
   def render(assigns) do
@@ -18,7 +18,7 @@ defmodule GrassFarmerWeb.Components.Zones do
         <% end %>
       </div>
       <div class="flex justify-end">
-        <button class="rounded-full text-2xl font-bold bg-green-200 px-3 py-1" phx-click="add_zone">
+        <button class="rounded-full text-2xl font-bold bg-green-200 px-3 py-1" phx-click="add_zone" phx-target={@myself}>
           +
         </button>
       </div>
@@ -31,7 +31,12 @@ defmodule GrassFarmerWeb.Components.Zones do
     ~H"""
     <div class={bg_color(@zone.status) <> " each flex hover:shadow-lg select-none p-2 rounded-md border-gray-300 border mb-1 hover:border-gray-500 cursor-pointer"}>
       <div class="left">
-        <%= if @zone.edit do %> Editing
+        <%= if @zone.edit == true do %>
+          <form action="#" phx-submit="update_zone">
+            <input type="text" name="zone_name" placeholder={@zone.name} />
+            <input type="text" name="zone_id" placeholder={@zone.id} value={@zone.id} class="w-12 text-center"/>
+            <input type="submit" value="update" class="text-blue-500 cursor-pointer pl-4">
+          </form>
         <% else %>
             <span class="header text-blue-600 font-semibold text-lb"><%= @zone.name %></span>
             <span class="text-gray-600 font-semibold text-sm"><%= @zone.id %></span>
@@ -50,6 +55,20 @@ defmodule GrassFarmerWeb.Components.Zones do
 
     </div>
     """
+  end
+
+  @impl true
+  def handle_event("add_zone", _params, socket) do
+    zone_max_id =
+      socket.assigns.zones
+      |> Enum.reduce(0, fn zone, acc -> max(zone.id, acc) end)
+
+    zones = socket.assigns.zones ++ [%Zone{id: zone_max_id + 1}]
+
+    PersistenceAdapter.new(%{set_name: "zones", configs: zones})
+    |> PersistenceAdapter.local_write
+
+    {:noreply, assign(socket, %{zones: zones})}
   end
 
   defp bg_color(status) do
