+++ lib/grass_farmer_web/live/home.ex
@@ -13,7 +13,6 @@ defmodule GrassFarmerWeb.Home do
       assign(socket,
         Loader.translate() |> Map.merge(%{time: NaiveDateTime.local_now()})
       )
-      |> IO.inspect(label: "new_socket")
 
     {:ok, new_socket}
   end
@@ -38,19 +37,19 @@ defmodule GrassFarmerWeb.Home do
     """
   end
 
+
   @impl true
-  def handle_event("add_zone", _params, socket) do
-    zone_max_id =
+  def handle_event("update_zone", params, socket) do
+    new_zones =
       socket.assigns.zones
-      |> Enum.reduce(0, fn zone, acc -> max(zone.id, acc) end)
+      |> Enum.map(fn zone -> if zone.id == (params["zone_id"] |> String.to_integer), do: %Zone{zone | name: params["zone_name"], id: params["zone_id"], edit: false}, else: zone end)
 
-    zones = socket.assigns.zones ++ [%Zone{id: zone_max_id + 1}]
+    PersistenceAdapter.new(%{set_name: "zones", configs: new_zones})
+     |> PersistenceAdapter.local_write
 
-    PersistenceAdapter.new(%{set_name: "zones", configs: zones})
-    |> PersistenceAdapter.local_write
-
-    {:noreply, assign(socket, %{zones: zones})}
+    {:noreply, assign(socket, %{zones: new_zones})}
   end
+
 
   @impl true
   def handle_event("edit_zone", %{"zone" => zone_id}, socket) do
@@ -63,7 +62,6 @@ defmodule GrassFarmerWeb.Home do
 
     {:noreply,
      assign(socket, %{zones: new_zones})}
-
   end
 
   @impl true
